When would wave support Laravel 9
Firstly, I would like to apologize for using an unconventional means to  make this discussion. But considering that discussions are turned off, I had to. I'm sorry.

I just would like to ask, is there any plan soon for wave to support Laravel 9? As i am not a pro member, I do not know if there are any steps/discussions being made to allow Wave to support Laravel 9.

The reason I ask is because I was planning to get involved in upgrading Wave to Laravel 9, and merge it as a pull request. But I want to be sure that there are no already set plans for that to prevent conflicts and prevent me from wasting my time.

Thank so you much for you anticipated reply. Keep up the good work!